✨ Apply memo API and modify time flow (#48)
