+++ src/routes/Memo.tsx
@@ -1,20 +1,56 @@
 import {useState, ChangeEvent} from 'react';
-import {useLocation, useNavigate} from 'react-router-dom';
+import {useNavigate} from 'react-router-dom';
 import * as C from '../styles/CommonStyle';
 import * as S from '../styles/MemoStyle';
 import TopBackXBar from '@/components/common/TopBackXBar';
 import ProgressBar from '@/components/request/ProgressBar';
 import {applyService} from '@/utils/requestApi';
 
-// onClick={() => handleNavLinkClick('/request/pay')}
-
 function Time() {
+  const navigate = useNavigate();
   const [inputData, setInputData] = useState<string>(''); // inputData의 타입을 문자열로 지정
 
   // 입력 값 변경 핸들러
   const handleInputChange = (e: ChangeEvent<HTMLTextAreaElement>) => {
     const value = e.target.value;
     setInputData(value);
+  };
+
+  const handleNavLinkClick = async (path: string): Promise<void> => {
+    try {
+      const childId = Number(localStorage.getItem('childId'));
+      const medicalCertificateId = Number(
+        localStorage.getItem('medicalCertificateId'),
+      );
+      const absenceCertificateId = Number(
+        localStorage.getItem('absenceCertificateId'),
+      );
+      const startDate = String(localStorage.getItem('startDate'));
+      const endDate = String(localStorage.getItem('endDate'));
+
+      // applyService 호출 후 applyId를 반환받기
+      const applyId = await applyService({
+        childId,
+        medicalCertificateId,
+        absenceCertificateId,
+        startDate,
+        endDate,
+        memo: inputData,
+      });
+
+      localStorage.removeItem('childId');
+      localStorage.removeItem('medicalCertificateId');
+      localStorage.removeItem('absenceCertificateId');
+      localStorage.removeItem('startDate');
+      localStorage.removeItem('endDate');
+
+      console.log('Navigating to next page with applyId:', applyId);
+
+      // applyId를 다음 페이지로 넘기기 query string으로
+      navigate(`${path}?applyId=${applyId}`);
+    } catch (error: any) {
+      console.error('Error submitting request:', error.message);
+    }
   };
 
   return (
@@ -41,7 +77,9 @@ function Time() {
                   />
                 </S.Container>
                 <S.FooterContainer>
-                  <S.Button>다음</S.Button>
+                  <S.Button onClick={() => handleNavLinkClick('/request/pay')}>
+                    다음
+                  </S.Button>
                 </S.FooterContainer>
               </S.PageSpace>
             </C.PageSpace>
